==> 1-OSM.ipynb <==
!pip install "package" --user

!pip install -U osmnx 

import geopandas as gpd
import pandas as pd
import osmnx as ox
import pickle
from shapely import LineString,MultiLineString
import numpy as np
import difflib

path=r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm"

iris_data=pd.read_excel('Y:\REFERENTIEL DATA\RP 2022 FRANCE\GEODATA - IRIS\\Geodata_IRIS_essentiels.xlsx')

contour_iris=gpd.read_file(r'Y:\REFERENTIEL DATA\RP 2022 FRANCE\CARTOGRAPHIE\Cartographie_France_IRIS_2019\CONTOURS-IRIS_2-1_SHP_LAMB93_FXX-2019\CONTOURS-IRIS.shp')

regions=["Auvergne-Rhône-Alpes",
    "Bourgogne-Franche-Comté",
    "Bretagne",
    "Centre-Val de Loire",
    "Corse",
    "Grand Est",
    'Hauts-de-France',
    "Ile-de-France",
    "Normandie",
    "Nouvelle-Aquitaine",
    "Occitanie",
    'Pays de la Loire',
    "Provence-Alpes-Côte d'Azur"]


lyon_metropole=["Lyon",
"Albigny-sur-Saône",
"Bron",
"Cailloux-sur-Fontaines",
"Caluire-et-Cuire",
"Champagne-au-Mont-d'Or",
"Charbonnières-les-Bains",
"Charly",
"Chassieu",
"Collonges-au-Mont-d'Or",
"Corbas",
"Couzon-au-Mont-d'Or",
"Craponne",
"Curis-au-Mont-d'Or",
"Dardilly",
"Décines-Charpieu",
"Écully",
"Feyzin",
"Fleurieu-sur-Saône",
"Fontaines-Saint-Martin",
"Fontaines-sur-Saône",
"Francheville",
"Genay",
"Givors",
"Grigny",
"Irigny",
"Jonage",
"Limonest",
"Lissieu",
"Marcy-l'Étoile",
"Meyzieu",
"Mions",
"Montanay",
"La Mulatière",
"Neuville-sur-Saône",
"Oullins",
"Pierre-Bénite",
"Poleymieux-au-Mont-d'Or",
"Quincieux",
"Rillieux-la-Pape",
"Rochetaillée-sur-Saône",
"Saint-Cyr-au-Mont-d'Or",
"Saint-Didier-au-Mont-d'Or",
"Sainte-Foy-lès-Lyon",
"Saint-Fons",
"Saint-Genis-Laval",
"Saint-Genis-les-Ollières",
"Saint-Germain-au-Mont-d'Or",
"Saint-Priest",
"Saint-Romain-au-Mont-d'Or",
"Sathonay-Camp",
"Sathonay-Village",
"Solaize",
"Tassin-la-Demi-Lune",
"La Tour-de-Salvagny",
"Vaulx-en-Velin",
"Vénissieux",
"Vernaison",
"Villeurbanne"]

list_departement=pd.read_excel(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\\Liste-Excel-des-departements-francais.xlsx")

list_departement=list_departement[list_departement["Région"].isin(regions)]

# Batiments

def get_buildings(place_name,region):
    tags = {"building":True }
    
    if place_name.lower()!="rhône": 
        try:
            buildings = ox.geometries_from_place(place_name+", France", tags)
        except:
            buildings = ox.geometries_from_place(place_name, tags)
    else:
        buildings = ox.geometries_from_place([place_name,"métropole de Lyon"], tags)
        
    selected_columns=["geometry" ,"building", "shop" , "building:levels","amenity","name"]
    buildings=buildings[selected_columns]
    #pickle.dump(buildings,open(path+"\\"+region.lower()+"\\"+place_name.lower()+"\\raw data\\buildings.pkl",'wb'))   
    return buildings

get_buildings('paris',"Ile de France")

# amenity (hopital , école ...)

def get_amenity(place_name,region):
    tags = {"amenity":True }
    if place_name.lower()!="rhône": 
        try:
            amenity = ox.geometries_from_place(place_name+", France", tags)
        except:
            amenity = ox.geometries_from_place(place_name, tags)
    else:
        amenity = ox.geometries_from_place([place_name,"métropole de Lyon"], tags)
    selected_columns=["geometry" ,"building", "shop" , "building:levels","amenity","name"]
    amenity=amenity[selected_columns]
    
    pickle.dump(amenity,open(path+"\\"+region.lower()+"\\"+place_name.lower()+"\\raw data\\amenity.pkl","wb"))

# Natural

def get_natural (place_name,region):
    tags = {"natural":True }   
    if place_name.lower()!="rhône": 
        try:
            natural = ox.geometries_from_place(place_name+", France", tags)
        except:
            natural = ox.geometries_from_place(place_name, tags)
    else:
        natural = ox.geometries_from_place([place_name,"métropole de Lyon"], tags)
        
    natural=natural[["geometry",'natural','name']]
    pickle.dump(natural,open(path+"\\"+region.lower()+"\\"+place_name.lower()+"\\raw data\\natural.pkl","wb"))

# Landuse

def get_landuse(place_name,region):
    tags = {"landuse":True } 
    if place_name.lower()!="rhône": 
        try:
            landuse = ox.geometries_from_place(place_name+", France", tags)
        except:
            landuse = ox.geometries_from_place(place_name, tags)
    else:
        landuse = ox.geometries_from_place([place_name,"métropole de Lyon"], tags)
    landuse=landuse[["geometry",'landuse','name']]
    pickle.dump(landuse,open(path+"\\"+region.lower()+"\\"+place_name.lower()+"\\raw data\\landuse.pkl","wb"))

# Shops (restaurant ,Vêtements...)

def get_shops (place_name,region):
    tags = {"shop":True }   
    if place_name.lower()!="rhône": 
        try:
            shops = ox.geometries_from_place(place_name+", France", tags)
        except:
            shops = ox.geometries_from_place(place_name, tags)
    else:
        shops = ox.geometries_from_place([place_name,"métropole de Lyon"], tags)
        
    selected_columns=["geometry" ,"building", "shop" ,"amenity","name", ]
    shops=shops[selected_columns]
    pickle.dump(shops,open(path+"\\"+region.lower()+"\\"+place_name.lower()+"\\raw data\\shops.pkl","wb"))

# Transport

def get_transport(place_name,region):
    tags = {"railway":["station",'tram_stop'],"station":"subway"}
    if place_name.lower()!="rhône": 
        try:
            public_transports = ox.geometries_from_place(place_name+", France", tags=tags)
        except:
            public_transports = ox.geometries_from_place(place_name, tags=tags)
    else:
        public_transports = ox.geometries_from_place([place_name,"métropole de Lyon"], tags)
    selected_columns=['geometry' ,'public_transport','name',"railway"]
    if "station" in public_transports.columns:
        selected_columns.append('station')
    
    public_transports=public_transports[selected_columns]
    pickle.dump(public_transports,open(path+"\\"+region.lower()+"\\"+place_name.lower()+"\\raw data\\public_transports.pkl","wb"))

# Rues


def get_highway(place_name,region):
    tags = {"highway":True }
    
    if place_name.lower()!="rhône":  
        try:
            highways = ox.geometries_from_place(place_name+", France", tags)
        except:
            highways = ox.geometries_from_place(place_name, tags)
    else:

        highways = ox.geometries_from_place([place_name,"métropole de Lyon"], tags)
    selected_columns=['geometry' ,'highway','name']
    highways=highways[selected_columns]
    pickle.dump(highways,open(path+"\\"+region.lower()+"\\"+place_name.lower()+"\\raw data\\highways.pkl","wb"))
    return highways

# Tourism

def get_tourism(place_name,region):
    tags = {"tourism":True }   
    if place_name.lower()!="rhône": 
        try:
            tourism = ox.geometries_from_place(place_name+", France", tags)
        except:
            tourism = ox.geometries_from_place(place_name, tags)
    else:
        tourism = ox.geometries_from_place([place_name,"métropole de Lyon"], tags)
    selected_columns=["name", "geometry","tourism","stars"]
    tourism=tourism[selected_columns]
    pickle.dump(tourism,open(path+"\\"+region.lower()+"\\"+place_name.lower()+"\\raw data\\tourism.pkl","wb"))

# Loisir

def get_leisure(place_name,region):
    tags = {"leisure":True }
    if place_name.lower()!="rhône": 
        try:
            leisure = ox.geometries_from_place(place_name+", France", tags)
        except:
            leisure = ox.geometries_from_place(place_name, tags)
    else:
        leisure = ox.geometries_from_place([place_name,"métropole de Lyon"], tags)
        
        
    selected_columns=["name", "geometry","leisure"]
    leisure=leisure[selected_columns]
    pickle.dump(leisure,open(path+"\\"+region.lower()+"\\"+place_name.lower()+"\\raw data\\leisure.pkl","wb"))

# For Loop

#creation des dossiers
from pathlib import Path

for i,row in list_departement.iterrows():

        Path(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm\\"+row['Région'].lower()+"\\"+row["Département"].lower()+"\\raw data").mkdir(parents=True, exist_ok=True)
        Path(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm\\"+row['Région'].lower()+"\\"+row["Département"].lower()+"\\processed data").mkdir(parents=True, exist_ok=True)
        Path(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm\\"+row['Région'].lower()+"\\"+row["Département"].lower()+"\\association").mkdir(parents=True, exist_ok=True)


for i,row in list_departement.iterrows():
    place_name=row['Département']
  

    '''bat=get_buildings(place_name,row["Région"])
    get_amenity(place_name,row["Région"])
    get_natural(place_name,row["Région"])
    get_landuse(place_name,row["Région"])
    get_shops(place_name,row["Région"])
    get_tourism(place_name,row["Région"])
    get_highway(place_name,row["Région"])'''
    get_transport(place_name,row["Région"])
    #get_leisure(place_name,row["Région"])


        

import os
index=[]
for i,f in enumerate(list_departement.apply(lambda x:x["Région"].lower()+"\\"+x['Département'].lower()+"\\raw data",axis=1).values):
    if len(os.listdir(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm\\"+f)) !=9:
        index.append(i)


err_dep=list_departement.iloc[index]
err_dep

# Traitement des données

import pickle 
import pandas as pd
import numpy as np

#logs and stats
logs={}

important_columns=["unclassified", "residential","service","path","unclassified", "pedestrian","primary" ,"secondary" ,"tertiary","living_street" ,"trunk" ,"motorway" ]
for i ,row in list_departement.iterrows():
    try:
        place_name=row["Département"].lower()
        path=r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm\\"+row['Région'].lower()+r"\\"+row["Département"].lower()

        #Batiments
        data=pickle.load(open(path+r"\\raw data\buildings.pkl","rb")).to_crs("3857")
        n=len(data)
        data=data[data.area>=30]
        pickle.dump(data,open(path+r"\\processed data\buildings.pkl","wb"))
        #logs
        logs[place_name]={"batiment":{

            "types":data['building'].value_counts(),
            "building:levels":data['building:levels'].isna().sum(),
            "size_raw":n,
            "size":len(data),
            "type_yes":len(data[data['building']=="yes"])}
        }




        #Rues
        highways=pickle.load(open(path+r"\\raw data\highways.pkl","rb")).to_crs("3857")
        n=len(highways)
        highways=highways[highways['highway'].isin(important_columns)]
        highways=highways[~highways['name'].isna()]
        highways=highways[highways['geometry'].apply (lambda x:isinstance(x,(LineString,MultiLineString)))]
        highways=highways[ (highways['highway']!="service")]
        highways=highways[highways.length>=50]
        pickle.dump(highways,open(path+r"\\processed data\highways.pkl","wb"))
        #logs
        logs[place_name].update({"highways":{

            "types":highways['highway'].value_counts(),
            "size_raw":n,
            "size":len(highways),
        }})

        #Tourisme
        tourism=pickle.load(open(path+r"\\raw data\tourism.pkl","rb")).to_crs("3857")
        n=len(tourism)
        tourism=tourism[tourism['tourism']!="information"]
        pickle.dump(tourism,open(path+r"\\processed data\tourism.pkl","wb"))
        #logs
        logs[place_name].update({"tourism":{

            "types":tourism['tourism'].value_counts(),
            "size_raw":n,
            "size":len(tourism),
        }})

        #Amenity
        amenity=pickle.load(open(path+r"\\raw data\amenity.pkl","rb")).to_crs("3857")
        n=len(amenity)
        leisure=pickle.load(open(path+r"\\raw data\leisure.pkl","rb")).to_crs("3857")

        amenity=amenity.drop(np.intersect1d(np.array(tourism.index),np.array(amenity.index)))
        amenity=amenity.drop(np.intersect1d(np.array(leisure.index),np.array(amenity.index)))
        pickle.dump(amenity,open(path+r"\\processed data\amenity.pkl","wb"))
        #logs
        logs[place_name].update({"amenity":{

            "types":amenity['amenity'].value_counts(),
            "size_raw":n,
            "size":len(amenity),
        }})
        logs[place_name].update({"leisure":{

            "types":leisure['leisure'].value_counts(),
            "size":len(amenity),
        }})


    except:
        print(place_name)



pickle.dump(logs,open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\\logs.pkl","wb"))
