==> 5-filtrer les données .ipynb <==
import geopandas as gpd
import pandas as pd
import osmnx as ox
import pickle
from shapely import LineString,MultiLineString
import numpy as np
import difflib

path=r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm"
iris_data=pd.read_excel('Y:\REFERENTIEL DATA\RP 2022 FRANCE\GEODATA - IRIS\\Geodata_IRIS_essentiels.xlsx')
contour_iris=gpd.read_file(r'Y:\REFERENTIEL DATA\RP 2022 FRANCE\CARTOGRAPHIE\Cartographie_France_IRIS_2019\CONTOURS-IRIS_2-1_SHP_LAMB93_FXX-2019\CONTOURS-IRIS.shp').to_crs("3857")

commune_selected=pickle.load(open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\villes\\villes.pkl","rb")).to_crs("3857")


list_departement=pd.read_excel(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\\Liste-Excel-des-departements-francais.xlsx")

display_data=[]
for i ,row in list_departement.iterrows():
    try:
        place_name=row["Département"].lower()
        region=row['Région'].lower()
        print(place_name)
        
        path=r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm\\"+row['Région'].lower()+r"\\"+row["Département"].lower()

        #Batiments
 
        data=pickle.load(open(path+r"\\processed data\buildings.pkl","rb")).to_crs("3857")
        
        l=[list(data[data.within(commune_selected.iloc[j][0])].index) for j in range(len(commune_selected))]
        flat_list = [item for sublist in l for item in sublist]
        if len(flat_list)==0:continue
            

        pickle.dump(data.loc[flat_list],open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm filtre\batiment\\"+place_name+r"_buildings.pkl","wb"))
        display_data.append(data.loc[flat_list])
        
        #rues
        
        highways=pickle.load(open(path+r"\\processed data\highways.pkl","rb")).to_crs("3857")
        l=[list(highways[highways.intersects(commune_selected.iloc[j][0])].index) for j in range(len(commune_selected))]
        flat_list = [item for sublist in l for item in sublist]
        pickle.dump(highways.loc[flat_list],open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm filtre\rue\\"+place_name+r"_highways.pkl","wb"))
        
        
        #tourisme
        tourism=pickle.load(open(path+r"\\processed data\tourism.pkl","rb")).to_crs("3857")
        l=[list(tourism[tourism.within(commune_selected.iloc[j][0])].index) for j in range(len(commune_selected))]
        flat_list = [item for sublist in l for item in sublist]
        pickle.dump(tourism.loc[flat_list],open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm filtre\tourisme\\"+place_name+r"_tourism.pkl","wb"))
        
        
        
        #Amenity
        amenity=pickle.load(open(path+r"\\processed data\amenity.pkl","rb")).to_crs("3857")
        l=[list(amenity[amenity.within(commune_selected.iloc[j][0])].index) for j in range(len(commune_selected))]
        flat_list = [item for sublist in l for item in sublist]
        pickle.dump(amenity.loc[flat_list],open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm filtre\amenity\\"+place_name+r"_amenity.pkl","wb"))

        #Loisire
        leisure=pickle.load(open(path+r"\\raw data\leisure.pkl","rb")).to_crs("3857")
        l=[list(leisure[leisure.within(commune_selected.iloc[j][0])].index) for j in range(len(commune_selected))]
        flat_list = [item for sublist in l for item in sublist]
        pickle.dump(leisure.loc[flat_list],open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm filtre\loisire\\"+place_name+r"_leisure.pkl","wb"))
    except:
        print('error')

display_data=[]
for i ,row in list_departement.iterrows():
    try:
        place_name=row["Département"].lower()
        region=row['Région'].lower()
        print(place_name)
        
        path=r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm\\"+row['Région'].lower()+r"\\"+row["Département"].lower()

        data=pickle.load(open(path+r"\\processed data\buildings.pkl","rb")).to_crs("3857")
        if len(flat_list)==0:continue

        
        #shops
        
        shops=pickle.load(open(path+r"\\raw data\shops.pkl","rb")).to_crs("3857")
        l=[list(shops[shops.intersects(commune_selected.iloc[j][0])].index) for j in range(len(commune_selected))]
        lat_list = [item for sublist in l for item in sublist]
        pickle.dump(shops.loc[flat_list],open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm filtre\magasin\\"+place_name+r"_shops.pkl","wb"))
        
        
        #GSA
        gsa=pickle.load(open(path+r"\\processed data\gsa_comp.pkl","rb")).to_crs("3857")
        l=[list(gsa[gsa.intersects(commune_selected.iloc[j][0])].index) for j in range(len(commune_selected))]
        flat_list = [item for sublist in l for item in sublist]
        pickle.dump(gsa.loc[flat_list],open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm filtre\GSA\\"+place_name+r"_gsa.pkl","wb"))
        
        #GSS
        gss=pickle.load(open(path+r"\\processed data\gss_comp.pkl","rb")).to_crs("3857")
        l=[list(gss[gss.intersects(commune_selected.iloc[j][0])].index) for j in range(len(commune_selected))]
        flat_list = [item for sublist in l for item in sublist]
        pickle.dump(gss.loc[flat_list],open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm filtre\GSS\\"+place_name+r"_gss.pkl","wb"))

    except:
        print('error')

display_data=[]
for i ,row in list_departement.iterrows():
    
        place_name=row["Département"].lower()
        region=row['Région'].lower()
        
     
        path=r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm\\"+row['Région'].lower()+r"\\"+row["Département"].lower()
        public_transports=pickle.load(open(path+r"\\raw data\public_transports.pkl","rb")).to_crs("3857")


        print(place_name)
        #public transport 
        
        
        l=[list(public_transports[public_transports.intersects(commune_selected.iloc[j][0])].index) for j in range(len(commune_selected))]
        flat_list = [item for sublist in l for item in sublist]
        pickle.dump(public_transports.loc[flat_list],open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm filtre\public_transports\\"+place_name+r".pkl","wb"))


d=pd.read_pickle(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm filtre\public_transports\\paris.pkl")
