==> 6-b-fill_level_na_ML.ipynb <==
import geopandas as gpd
import osmnx as ox
import xgboost as xgb
import pickle
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor,GradientBoostingRegressor
import time 
import warnings
import time
import pandas as pd
import matplotlib.pyplot as plt

path=r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm filtre\batiment"
list_departement=pd.read_excel(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\\Liste-Excel-des-departements-francais.xlsx")

iris_data=pd.read_excel("Y:\REFERENTIEL DATA\RP 2022 FRANCE\GEODATA - IRIS\Geodata_IRIS_essentiels.xlsx")
contour_iris=gpd.read_file(r'Y:\REFERENTIEL DATA\RP 2022 FRANCE\CARTOGRAPHIE\Cartographie_France_IRIS_2019\CONTOURS-IRIS_2-1_SHP_LAMB93_FXX-2019\CONTOURS-IRIS.shp').to_crs('EPSG:3857')

building_no_na=building[~building['building:levels'].isna()]
building_na=building[building['building:levels'].isna()]

#get contours of the departement only 
dep_contour_iris=contour_iris[contour_iris['INSEE_COM'].isin([str(i) for i in range(dep*1000,dep*1000+1000)])]

a=gpd.sjoin(building,dep_contour_iris,predicate="intersects",how="left")
a = a[~a.index.duplicated(keep='first')].reset_index()

iris_building_id=a[["index_right",'element_type',"osmid"]].copy()
iris_building_id=iris_building_id.set_index(['element_type',"osmid"])

a=gpd.GeoDataFrame(
    a[['element_type',"osmid","CODE_IRIS"]].merge(contour_iris[['CODE_IRIS',"geometry"]]),

).set_geometry("geometry").set_crs('EPSG:3857')

xx=gpd.sjoin(a,dep_contour_iris,predicate="touches")

nearby_buildings=xx[["osmid",'element_type',"CODE_IRIS_left","index_right"]].groupby(["element_type",'osmid'])['index_right'].apply(list)

nearby_buildings=nearby_buildings.to_dict()

# ML

sample=building_no_na.sample(frac=1,random_state=123)

def f (x):
    if x!=x:
        return -100
    else:
        return float(x.split(';')[0])

data_no=None
#pour l'affichage
j=0

start=time.time()
warnings.filterwarnings("ignore")
k=10
#y_true
building_level=[]

print('number of rows :'+str(len(sample)))
for i ,row in sample.iterrows():
    if j%1000==0:
        print(j)
    building_level.append(row['building:levels'])
        
    #get the distance between all buildings and the current buildingi
    ## get the com nearby of the building i
    if i in nearby_buildings.keys() and nearby_buildings[i][0] is not None:
        current_iris_index=iris_building_id.loc[i]['index_right']
        iris_ids=list(nearby_buildings[i])
        iris_ids.append(current_iris_index)
        indices=list(iris_building_id[iris_building_id['index_right'].isin(iris_ids)].index)
        
    else:
        #indices=building.index
        indices=list(building_no_na.index)
        
    
    #considerer que les batiment dont on connait leurs nombre d'étages

    indices = building_no_na.index.intersection(indices)
    distance=building_no_na.loc[indices].distance(row['geometry']).sort_values()
    
    #remove the current building from the distance dataframe (distance will be 0 ofc)
    distance=distance.drop(i)

    #take the k nearest neighboors
    a=building_no_na.loc[distance.index].iloc[:k]

    #buildings nearby
    areas=a.area.to_numpy()
    distances=distance.iloc[:k].to_numpy()
    types=a['building'].to_numpy()
    levels=np.vectorize (f )(a['building:levels'].to_numpy())
    
    if len(areas)<k:
        areas=np.append(areas,[np.nan for _ in range(k-len(a))])
    if len(distances)<k:distances=np.append(distances,[np.nan for _ in range(k-len(a))])
    if len(types)<k:types=np.append(types,[np.nan for _ in range(k-len(a))])
    if len(levels)<k:levels=np.append(levels,[np.nan for _ in range(k-len(a))])
    
    #information about the iris
    iris=contour_iris[contour_iris.intersects(row['geometry'])].iloc[0]
    if iris['TYP_IRIS'] ==iris['TYP_IRIS']:
        iris_type=np.array([iris['TYP_IRIS']])
    else:
        iris_type=None
        
    iris_area=np.array([iris['geometry'].area])
    
    #information about the building itself
    type_building=np.array([row['building']])
    area_building=np.array([row['geometry'].area])
    
    #densité 
    area=iris["geometry"].area/10**6
    iris=iris_data[iris_data['CODE_IRIS']==iris["CODE_IRIS"]]
    pop=iris['P17_POP'].iloc[0]
    densite=[pop/area]

    

    if data_no is None:
        
        data_no=np.concatenate([type_building,area_building,iris_type,iris_area,densite,areas,types,levels,distances])
    else:
        
        data_no=np.vstack([data_no,np.concatenate([type_building,area_building,iris_type,iris_area,densite,areas,types,levels,distances])])

        
            
    j+=1
end=time.time()
print(end-start)

#create a dataframe

columns=["type_building","area_building","iris_type","iris_area","iris_densite"]
for j in ["areas","types","levels","distances"]:
    for i in range(k):
        columns.append(j+str(i))

data_no=pd.DataFrame(data_no,columns=columns)
data_no.head()

#building_level is the true label Y
building_level=np.array([float(i.split(';')[0]) for i in building_level])
building_level=building_level[~data_no.isna().any(axis=1)]
building_level.shape

#convert columns from object to int
for c in data_no:
    if "area" in c or "levels" in c or "distances" in c or "iris_densite" in c:
        data_no[c]=data_no[c].astype(np.int)

list_encoders=[]
from sklearn.preprocessing import OrdinalEncoder
for c in data_no.columns:
    if data_no[c].dtype=="O":
        print(c)
        enc = OrdinalEncoder()
        data_no[c]=enc.fit_transform(data_no[c].to_numpy().reshape(-1,1) ).squeeze().astype(np.int)
        list_encoders.append(enc)
        

X_train, X_test, y_train, y_test =train_test_split(data_no,building_level,random_state=123)

# XGBoost

dtrain = xgb.DMatrix(X_train, label=y_train)
dtest = xgb.DMatrix(X_test, label=y_test)

param = {'max_depth': 5, 'eta': 0.1, 'objective': 'reg:squarederror',"gamma":2,"lambda":4,"verbosity":0}
param['nthread'] = 4
param['eval_metric'] = ["mae"]
param['verbosity']=0
evallist = [(dtrain, 'train'), (dtest, 'eval')]
num_round = 500

start=time.time()
model = xgb.train(param, dtrain, num_round)
### ADDD Other metrics STD and mediane absolute error
end=time.time()
ypred = model.predict(dtest)
print('The error is :'+str(np.sum(np.abs(model.predict(dtest).round()-y_test))/len(y_test)))
print('Train error is :'+str(np.sum(np.abs(model.predict(dtrain).round()-y_train))/len(y_train)))

print("Std train error:",np.std(np.abs(model.predict(dtrain).round()-y_train)))
print("Std test error:",np.std(np.abs(model.predict(dtest).round()-y_test)))

print("Median Abs error ",np.std(np.abs(model.predict(dtest).round()-y_test)))
print('Train time '+str(end-start))

pickle.dump(model, open(place_name.lower()+".model.pkl","wb"))

# Stats

intervals=[5,10,15,20]

ypred=model.predict(dtest)

for i,el in enumerate(intervals):
    
    if i ==0:
        a="0"
        b=str(el)
        mask=y_test<=el
    elif i ==len(intervals)-1:
        a="+infini"
        mask=y_test>el
    else:
        a=str(intervals[i-1])
        b=str(el)
        mask=np.logical_and(y_test<=el , y_test>intervals[i-1])

    print('For intervale ['+a+","+b+"] The error is :"+str(np.sum(np.abs(ypred[mask].round()-y_test[mask]))/len(y_test[mask])))
        

da=pd.DataFrame([],columns=["level","error",'frequency'])
for i,el in enumerate(pd.Series(y_test).value_counts().index):
    mask=y_test==el
    da.loc[len(da)]=[el,np.sum(np.abs(ypred[mask].round()-y_test[mask]))/len(y_test[mask]),pd.Series(y_test).value_counts()[el]]

da.sort_values('error').head(10)

da.sort_values('error').corr()

**Conclusion**
L'erreur est trés bas pour lorsque le nombre d'étage est dans l'intervale [5-7] qui correpond au nombre d'étage moyen à Paris.
Les erreurs apparaissent pour les buildings ayant un trés grand nombre d'étage.

# Prediction des valeurs manquantes

data_no_pred=None
j=0
import warnings
warnings.filterwarnings("ignore")
k=10
#the value to predict
building_level=[]
print('number of rows :'+str(len(building_na)))
for i ,row in building_na.iterrows():
    if j%1000==0:
        print(j)

        
    #get the distance between all buildings and our building
    ## get com nearby of the building
    if i in nearby_buildings.keys() and nearby_buildings[i][0] is not None:
        current_iris_index=iris_building_id.loc[i]['index_right']
        iris_ids=list(nearby_buildings[i])
        iris_ids.append(current_iris_index)
        indices=list(iris_building_id[iris_building_id['index_right'].isin(iris_ids)].index)
        
    else:
        #indices=building.index
        indices=list(building_no_na.index)
    
    #considerer que les batiment dont on connait leurs nombre d'étages

    indices = building_no_na.index.intersection(indices)
    distance=building_no_na.loc[indices].distance(row['geometry']).sort_values()


    #take the k nearest neighboors
    a=building_no_na.loc[distance.index].iloc[:k]
    
    
    
    
    #buildings nearby
    areas=a.area.to_numpy()
    distances=distance.iloc[:k].to_numpy()
    types=a['building'].to_numpy()
    levels=np.vectorize (f )(a['building:levels'].to_numpy())
    if len(areas)<k:
        areas=np.append(areas,[np.nan for _ in range(k-len(a))])
    if len(distances)<k:distances=np.append(distances,[np.nan for _ in range(k-len(a))])
    if len(types)<k:types=np.append(types,[np.nan for _ in range(k-len(a))])
    if len(levels)<k:levels=np.append(levels,[np.nan for _ in range(k-len(a))])
    
    #information about the iris
    iris=contour_iris[contour_iris.intersects(row['geometry'])].iloc[0]
    if iris['TYP_IRIS'] ==iris['TYP_IRIS']:
        iris_type=np.array([iris['TYP_IRIS']])
    else:
        iris_type=None
        
    iris_area=np.array([iris['geometry'].area])
    
    #densité 
    area=iris["geometry"].area/10**6
    iris=iris_data[iris_data['CODE_IRIS']==iris["CODE_IRIS"]]
    pop=iris['P17_POP'].iloc[0]
    densite=[pop/area]

    #information about the building itself
    type_building=np.array([row['building']])
    area_building=np.array([row['geometry'].area])

    if data_no_pred is None:
        
        data_no_pred=np.concatenate([type_building,area_building,iris_type,iris_area,densite,areas,types,levels,distances])
    else:
    
        data_no_pred=np.vstack([data_no_pred,np.concatenate([type_building,area_building,iris_type,iris_area,densite,areas,types,levels,distances])])

        
            
    j+=1
    

#create a dataframe
columns=["type_building","area_building","iris_type","iris_area","iris_densite"]
for j in ["areas","types","levels","distances"]:
    for i in range(k):
        columns.append(j+str(i))

data_no_pred=pd.DataFrame(data_no_pred,columns=columns)
data_no_pred.head()

data_no_pred_backup=data_no_pred.copy()

#convert columns from object to int
for c in data_no_pred:
    if "area" in c or "levels" in c or "distances" in c or "iris_densite" in c:
        data_no_pred[c]=data_no_pred[c].astype(np.int)

i=0
for c in data_no_pred.columns:
    if data_no_pred[c].dtype=="O":

        
        enc=list_encoders[i]
        enc.set_params(encoded_missing_value=-1,handle_unknown="use_encoded_value")

        data_no_pred[c]=enc.transform(data_no_pred[c].to_numpy().reshape(-1,1) ).squeeze().astype(np.int)
        i+=1
        

data_no_pred=xgb.DMatrix(data_no_pred)

prediction=model.predict(data_no_pred).round()

pd.Series(prediction).value_counts()



j=0
for i , row in building.loc[building['building:levels'].isna()].iterrows():
    building.loc[i,"building:levels"]=str(prediction[j])
    j+=1

pickle.dump(building,open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm\\"+place_name.lower()+"\\processed data\\building_level_filled.pkl",'wb'))
