==> 10-a- modele de score (JP).ipynb <==
import geopandas as gpd
import pandas as pd
import osmnx as ox
import pickle
from shapely import LineString,MultiLineString
from shapely import geometry, ops
import numpy as np

data=pickle.load(open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm filtre\batiments_levels_types\\essonne_buildings.pkl",'rb'))

data=data[data.intersects(contour_iris[contour_iris.INSEE_COM=="91377"].unary_union)]

data.explore('building').save(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm filtre\batiments_levels_types\\map.html")

iris_data=pd.read_excel(r"Y:\REFERENTIEL DATA\RP 2022 FRANCE\GEODATA - IRIS\\Geodata_IRIS_essentiels.xlsx")
contour_iris=gpd.read_file(r'Y:\REFERENTIEL DATA\RP 2022 FRANCE\CARTOGRAPHIE\Cartographie_France_IRIS_2019\CONTOURS-IRIS_2-1_SHP_LAMB93_FXX-2019\CONTOURS-IRIS.shp').to_crs("3857")

path=r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\osm filtre\\"
list_departement=pd.read_excel(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\\Liste-Excel-des-departements-francais.xlsx")

final_result=None
for j ,row in list_departement.iterrows():
    try:
        place_name=row['Département']
        result=pickle.load(open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\04_données clients (source)\finale data\data\\"+place_name.lower()+".pkl","rb"))
    except:
        continue
    if final_result is None:
        final_result=result
    else:
        final_result=pd.concat([final_result,result])

final_result=gpd.GeoDataFrame(final_result, crs="EPSG:3857", geometry="geometry")

cols=['C17_MEN', 'C17_MEN_CS1', 'C17_MEN_CS2', 'C17_MEN_CS3', 'C17_MEN_CS4',
       'C17_MEN_CS5', 'C17_MEN_CS6', 'C17_MEN_CS7', 'C17_MEN_CS8', 'P17_POP',
       'P17_POP0002', 'P17_POP0305', 'P17_POP0610', 'P17_POP1117',
       'P17_POP1824', 'P17_POP2539', 'P17_POP4054', 'P17_POP5564',
       'P17_POP6579', 'P17_POP80P', 'P17_POP15P', 'Attracteur_tourisme',
       'metro', 'tram', 'transport',  'GSA',
       'Sport', 'Equipement maison', 'Equipement Personne', 'Culture Loisir',
       'Mobilité', 'Culture loisir -loisire', 'Sport -loisire',
       'Culture loisir -amenity', 'restauration', 'Soin Personne',
       'Soin de la personne et Médical -amenity', 'Alimentaire',
       'Service public', "pt d'intérert", 'Medical', 'tous -amenity',
       'Surface_de_vente_GSA', 'Attracteur pour tous -loisire']

pd.set_option('display.max_columns', 500)

# Transformation en qualitatif

## Varibale de population

final_result.describe()

copy_result=final_result.copy()

for c in ['C17_MEN', 'C17_MEN_CS2', 'C17_MEN_CS3', 'C17_MEN_CS4',
       'C17_MEN_CS5', 'C17_MEN_CS6', 'C17_MEN_CS7', 'C17_MEN_CS8', 'P17_POP',
       'P17_POP0002', 'P17_POP0305', 'P17_POP0610', 'P17_POP1117',
       'P17_POP1824', 'P17_POP2539', 'P17_POP4054', 'P17_POP5564',
       'P17_POP6579', 'P17_POP80P', 'P17_POP15P']:
    
    final_result[c]=pd.qcut(final_result[c],10,labels=False)+1

final_result["C17_MEN_CS1"]=pd.cut(final_result["C17_MEN_CS1"],10,labels=False)+1

## Le reste

#Attracteur_tourisme
final_result['Attracteur_tourisme']=final_result['Attracteur_tourisme'].clip(lower=1, upper=10)


#metro
final_result.loc[(final_result['metro']>2).values,"metro"]=10
final_result[['metro']]=final_result[['metro']].replace({ 1: 8, 2: 9,0:1})

#tram
final_result.loc[(final_result['tram']>2).values,"tram"]=10
final_result[['tram']]=final_result[['tram']].replace({ 1: 8, 2: 8,2:9,0:1})

#transport
final_result.loc[(final_result['transport']>2).values,"transport"]=10
final_result[['transport']]=final_result[['transport']].replace({ 1: 8, 2: 9,0:1})

#GSA
final_result.loc[(final_result['GSA']>2).values,"GSA"]=10
final_result[['GSA']]=final_result[['GSA']].replace({ 1: 8, 2: 9,0:1})

#Sport
final_result.loc[(final_result['Sport']>2).values,"Sport"]=10
final_result[['Sport']]=final_result[['Sport']].replace({ 1: 8, 2: 9,0:1})

#Equipement maison
final_result.loc[(final_result['Equipement maison']>2).values,"Equipement maison"]=10
#final_result[['Equipement maison']]=final_result[['Equipement maison']].replace({ 1: 6, 2: 7,3:8,4:9,0:1})
final_result[['Equipement maison']]=final_result[['Equipement maison']].replace({ 1: 8, 2: 9,0:1})

#Equipement Personne
final_result.loc[(final_result['Equipement Personne']>2).values,"Equipement Personne"]=10
final_result[['Equipement Personne']]=final_result[['Equipement Personne']].replace({ 1: 8, 2: 9,0:1})
#.replace({ 1: 5, 2: 6,3:7,4:8,5:9,0:1})

#Culture Loisir
final_result.loc[(final_result['Culture Loisir']>2).values,"Culture Loisir"]=10
final_result[['Culture Loisir']]=final_result[['Culture Loisir']].replace({ 1: 8, 2: 9,0:1})
#.replace({ 1: 5, 2: 6,3:7,4:8,5:9,0:1})

#Mobilité
final_result[['Mobilité']]=final_result[['Mobilité']].clip(1,10)

#Culture loisir -loisire
final_result.loc[(final_result['Culture loisir -loisire']>2).values,"Culture loisir -loisire"]=10
final_result[['Culture loisir -loisire']]=final_result[['Culture loisir -loisire']].replace({ 1: 7, 2: 8,3:9,0:1})

#Sport -loisire
final_result.loc[(final_result['Sport -loisire']>3).values,"Sport -loisire"]=10
final_result[['Sport -loisire']]=final_result[['Sport -loisire']].replace({ 1: 7, 2: 8,3:9,0:1})

#Culture loisir -amenity
final_result.loc[(final_result['Culture loisir -amenity']>3).values,"Culture loisir -amenity"]=10
final_result[['Culture loisir -amenity']]=final_result[['Culture loisir -amenity']].replace({ 1: 7, 2: 8,3:9,0:1})

#restauration
final_result.loc[(final_result['restauration']>2).values,"restauration"]=10
final_result[['restauration']]=final_result[['restauration']].replace({ 1: 8, 2: 9,0:1})
#.replace({ 1: 7, 2: 8,3:9,0:1})

#Soin Personne
final_result.loc[(final_result['Soin Personne']>2).values,"Soin Personne"]=10
final_result[['Soin Personne']]=final_result[['Soin Personne']].replace({ 1: 8, 2: 9,0:1})
#.replace({ 1: 5, 2: 6,3:7,4:8,5:9,0:1})

#Soin Personne amenity
final_result.loc[(final_result['Soin de la personne et Médical -amenity']>5).values,"Soin de la personne et Médical -amenity"]=10
final_result[['Soin de la personne et Médical -amenity']]=final_result[['Soin de la personne et Médical -amenity']].replace({ 1: 5, 2: 6,3:7,4:8,5:9,0:1})

#Alimentaire
final_result.loc[(final_result['Alimentaire']>5).values,"Alimentaire"]=10
final_result[['Alimentaire']]=final_result[['Alimentaire']].replace({ 1: 5, 2: 6,3:7,4:8,5:9,0:1})

#Service public
final_result.loc[(final_result['Service public']>2).values,"Service public"]=10
final_result[['Service public']]=final_result[['Service public']].replace({ 1: 8, 2: 9,0:1})
#.replace({ 1: 5, 2: 6,3:7,4:8,5:9,0:1})

#pt d'intérert
final_result.loc[(final_result["pt d'intérert"]>2).values,"pt d'intérert"]=10
final_result[["pt d'intérert"]]=final_result[["pt d'intérert"]].replace({ 1: 8, 2: 9,0:1})
#.replace({ 1: 5, 2: 6,3:7,4:8,5:9,0:1})

#Medical
final_result.loc[(final_result["Medical"]>2).values,"Medical"]=10
final_result[["Medical"]]=final_result[["Medical"]].replace({ 1: 8, 2: 9,0:1})
#.replace({ 1: 7, 2: 8,3:9,0:1})

#tous -amenity
final_result.loc[(final_result["tous -amenity"]>5).values,"tous -amenity"]=10
final_result[["tous -amenity"]]=final_result[["tous -amenity"]].replace({ 1: 5, 2: 6,3:7,4:8,5:9,0:1})

#Attracteur pour tous -loisire
final_result.loc[(final_result["Attracteur pour tous -loisire"]>5).values,"Attracteur pour tous -loisire"]=10
final_result[["Attracteur pour tous -loisire"]]=final_result[["Attracteur pour tous -loisire"]].replace({ 1: 5, 2: 6,3:7,4:8,5:9,0:1})

#Surface_de_vente_GSA
final_result.loc[(final_result["Surface_de_vente_GSA"]<=0).values,"Surface_de_vente_GSA"]=1
final_result.loc[((final_result["Surface_de_vente_GSA"]<=1000) & (final_result["Surface_de_vente_GSA"]>0) ).values,"Surface_de_vente_GSA"]=7
final_result.loc[((final_result["Surface_de_vente_GSA"]<=5000) & (final_result["Surface_de_vente_GSA"]>1000) ).values,"Surface_de_vente_GSA"]=8
final_result.loc[((final_result["Surface_de_vente_GSA"]<=10000) & (final_result["Surface_de_vente_GSA"]>5000) ).values,"Surface_de_vente_GSA"]=9
final_result.loc[(final_result["Surface_de_vente_GSA"]>10000) .values,"Surface_de_vente_GSA"]=10

matrice_pond=pd.read_excel(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\\Données_OSM_métie1.xlsx",sheet_name="matrice de pondération")
matrice_pond=matrice_pond.iloc[1:,1:]
matrice_pond.columns=matrice_pond.iloc[0,:].values
matrice_pond=matrice_pond.iloc[1:,:].fillna(0)
matrice_pond=matrice_pond.set_index(' ')
matrice_pond=matrice_pond.drop(0)

copy=final_result.copy()

scores=(copy[matrice_pond.index].values@matrice_pond/100)

'''from sklearn.preprocessing import MinMaxScaler
scaler=MinMaxScaler()
scores.loc[:,:]=scaler.fit_transform(scores)*10'''

scores=scores.round(2)

final_result.loc[:,["score "+i for i in  scores.columns]]=scores.values
final_result.loc[:,"Score Globale"]=scores.mean(axis=1).values

pickle.dump(final_result,open(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\\scores_données.pkl","wb"))
final_result.to_excel(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\\scores_données.xlsx")

cols_display=['P17_POP',"Attracteur_tourisme" , "metro",  "tram","transport","GSA",  "Sport",
             
             "Equipement maison",  "Equipement Personne",
              "Culture Loisir" , "Culture loisir -loisire",  

                        "Sport -loisire" , "Alimentaire" , "Mobilité",  

                       'Soin de la personne et Médical -amenity' , 

                        "pt d'intérert",  "Medical",  "restauration" , "Soin Personne" ,  

                        "Service public"  ,"tous -amenity"  ,

                        "Culture loisir -amenity",  


                        "Attracteur pour tous -loisire",  "Surface_de_vente_GSA","geometry","name" ,"Score Globale" 
              
             ]
cols_display.extend(["score "+i for i in  scores.columns])

final_result=copy_result.copy()
final_result.loc[:,["score "+i for i in  scores.columns]]=scores.values
final_result.loc[:,"Score Globale"]=scores.mean(axis=1).values

for c in scores:
    final_result[cols_display].iloc[scores[scores[c]>=7].index].explore("score "+c,cmap="Reds").save(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\cartes_scores\score 7\\"+c+"_score.html")

for c in scores:
    final_result[cols_display].iloc[scores[scores[c]>=5].index].explore("score "+c).save(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\cartes_scores\score 5\\"+c+"_score.html")

final_result.loc[((scores>=7).all(axis=1)).values].explore(color="black").save(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\\best.html")

final_result[final_result['Score Globale']>=7].explore('Score Globale').save(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\\best_scores_globale.html")

final_result[final_result.name=="Rue du Commerce"].loc[:,cols_display].explore().save(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\cartes_scores\\rue du commerce.html")

final_result[final_result.name=="Rue des Blanchers"].loc[:,cols_display].explore().save(r"Y:\RECHERCHE ET DEV\10_Modèle d'attractivite des rues\cartes_scores\\rue des blanchers.html")


final_result.loc[((scores>=7).all(axis=1)).values].shape

(final_result['Score Globale']>=7).sum()
